requirements.txt
Flask
flask-cors
gunicorn
psycopg2-binary
werkzeug